refactor: 각 감정 초기화 부분은 Emotions으로 이동
